Refactor is_valid method and update health route
The `is_valid` method was refactored, being added to `Huawei.py` and removed from `Metrics.py`. In `app.py`, the reference to `is_valid` in the '/health' route was updated to call `is_valid` properly. It enables health check endpoint to return "OK" if metrics are valid and "NOK" if they are not.